chores(Presto): move `PrestoBaseEngineSpec` to `presto.py`
